Added SimpleCache 2/3 compatible DevNullCache implementation
